+++ services/src/floodlight-service/floodlight-modules/src/main/java/org/openkilda/floodlight/command/bfd/BfdSessionCommand.java
@@ -78,7 +78,7 @@ abstract class BfdSessionCommand extends BfdCommand {
         OFActionNoviflowBfdStart bfdStartAction = ofFactory.actions().buildNoviflowBfdStart()
                 .setPortNo(bfdSession.getLogicalPortNumber())
                 .setMyDisc(bfdSession.getDiscriminator())
-                .setInterval(bfdSession.getIntervalMs() * 1000)
+                .setInterval(bfdSession.getIntervalMs() * 1000L)
                 .setMultiplier(bfdSession.getMultiplier())
                 .setKeepAliveTimeout(((short) (bfdSession.isKeepOverDisconnect() ? 1 : 0)))
                 .build();