+++ services/wfm/src/main/java/org/openkilda/wfm/topology/nbworker/NbWorkerTopology.java
@@ -19,10 +19,17 @@ import org.openkilda.pce.PathComputerConfig;
 import org.openkilda.persistence.PersistenceManager;
 import org.openkilda.persistence.spi.PersistenceProvider;
 import org.openkilda.wfm.LaunchEnvironment;
+import org.openkilda.wfm.share.flow.resources.FlowResourcesConfig;
+import org.openkilda.wfm.share.hubandspoke.CoordinatorBolt;
+import org.openkilda.wfm.share.hubandspoke.CoordinatorSpout;
+import org.openkilda.wfm.share.hubandspoke.HubBolt;
+import org.openkilda.wfm.share.hubandspoke.WorkerBolt;
 import org.openkilda.wfm.topology.AbstractTopology;
 import org.openkilda.wfm.topology.nbworker.bolts.DiscoveryEncoderBolt;
 import org.openkilda.wfm.topology.nbworker.bolts.FeatureTogglesBolt;
+import org.openkilda.wfm.topology.nbworker.bolts.FlowMeterModifyHubBolt;
 import org.openkilda.wfm.topology.nbworker.bolts.FlowOperationsBolt;
+import org.openkilda.wfm.topology.nbworker.bolts.FlowValidationHubBolt;
 import org.openkilda.wfm.topology.nbworker.bolts.HistoryOperationsBolt;
 import org.openkilda.wfm.topology.nbworker.bolts.KildaConfigurationBolt;
 import org.openkilda.wfm.topology.nbworker.bolts.LinkOperationsBolt;
@@ -30,13 +37,17 @@ import org.openkilda.wfm.topology.nbworker.bolts.MessageEncoder;
 import org.openkilda.wfm.topology.nbworker.bolts.PathsBolt;
 import org.openkilda.wfm.topology.nbworker.bolts.ResponseSplitterBolt;
 import org.openkilda.wfm.topology.nbworker.bolts.RouterBolt;
+import org.openkilda.wfm.topology.nbworker.bolts.SpeakerWorkerBolt;
 import org.openkilda.wfm.topology.nbworker.bolts.SwitchOperationsBolt;
-import org.openkilda.wfm.topology.nbworker.bolts.SwitchValidationsBolt;
+import org.openkilda.wfm.topology.utils.MessageKafkaTranslator;
 
 import org.apache.storm.generated.StormTopology;
 import org.apache.storm.kafka.bolt.KafkaBolt;
 import org.apache.storm.kafka.spout.KafkaSpout;
 import org.apache.storm.topology.TopologyBuilder;
+import org.apache.storm.tuple.Fields;
+
+import java.util.concurrent.TimeUnit;
 
 /**
  *  Storm topology to read data from database.
@@ -45,7 +56,6 @@ import org.apache.storm.topology.TopologyBuilder;
  *                                     | ---> links-operations-bolt      ---> |   \                       |
  *                                     | ---> flows-operations-bolt      ---> |  \ \                      |
  *                                     | ---> switches-operations-bolt   ---> | \ \ \                     |
- *                                     | ---> switch-validations-bolt    ---> |\ \ \ \                    |
  *                                     | ---> history-operations-bolt    ---> | \ \ \ \                   |
  *                                     | --->                   message-encoder-bolt (in error case) ---> |
  *
@@ -63,7 +73,6 @@ public class NbWorkerTopology extends AbstractTopology<NbWorkerTopologyConfig> {
     private static final String FEATURE_TOGGLES_BOLT_NAME = "feature-toggles-bolt";
     private static final String KILDA_CONFIG_BOLT_NAME = "kilda-config-bolt";
     private static final String PATHS_BOLT_NAME = "paths-bolt";
-    private static final String SWITCH_VALIDATIONS_BOLT_NAME = "switch-validations-bolt";
     private static final String MESSAGE_ENCODER_BOLT_NAME = "message-encoder-bolt";
     private static final String DISCOVERY_ENCODER_BOLT_NAME = "discovery-encoder-bolt";
     private static final String SPLITTER_BOLT_NAME = "response-splitter-bolt";
@@ -73,6 +82,11 @@ public class NbWorkerTopology extends AbstractTopology<NbWorkerTopologyConfig> {
     private static final String DISCO_KAFKA_BOLT_NAME = "disco-kafka-bolt";
     private static final String HISTORY_BOLT_NAME = "history-operations-bolt";
     private static final String NB_SPOUT_ID = "nb-spout";
+    private static final String SPEAKER_KAFKA_BOLT = "speaker-bolt";
+    private static final String VALIDATION_WORKER_BOLT = "validation." + SpeakerWorkerBolt.ID;
+    private static final String METER_MODIFY_WORKER_BOLT = "meter.modify." + SpeakerWorkerBolt.ID;
+
+    private static final Fields FIELDS_KEY = new Fields(MessageKafkaTranslator.FIELD_ID_KEY);
 
     public NbWorkerTopology(LaunchEnvironment env) {
         super(env, NbWorkerTopologyConfig.class);
@@ -86,6 +100,14 @@ public class NbWorkerTopology extends AbstractTopology<NbWorkerTopologyConfig> {
 
         final Integer parallelism = topologyConfig.getParallelism();
 
+        tb.setSpout(CoordinatorSpout.ID, new CoordinatorSpout());
+        tb.setBolt(CoordinatorBolt.ID, new CoordinatorBolt())
+                .allGrouping(CoordinatorSpout.ID)
+                .fieldsGrouping(FlowValidationHubBolt.ID, CoordinatorBolt.INCOME_STREAM, FIELDS_KEY)
+                .fieldsGrouping(VALIDATION_WORKER_BOLT, CoordinatorBolt.INCOME_STREAM, FIELDS_KEY)
+                .fieldsGrouping(FlowMeterModifyHubBolt.ID, CoordinatorBolt.INCOME_STREAM, FIELDS_KEY)
+                .fieldsGrouping(METER_MODIFY_WORKER_BOLT, CoordinatorBolt.INCOME_STREAM, FIELDS_KEY);
+
         KafkaSpout kafkaSpout = buildKafkaSpout(topologyConfig.getKafkaTopoNbTopic(), NB_SPOUT_ID);
         tb.setSpout(NB_SPOUT_ID, kafkaSpout, parallelism);
 
@@ -96,6 +118,53 @@ public class NbWorkerTopology extends AbstractTopology<NbWorkerTopologyConfig> {
         PersistenceManager persistenceManager =
                 PersistenceProvider.getInstance().createPersistenceManager(configurationProvider);
         PathComputerConfig pathComputerConfig = configurationProvider.getConfiguration(PathComputerConfig.class);
+
+        FlowResourcesConfig flowResourcesConfig = configurationProvider.getConfiguration(FlowResourcesConfig.class);
+
+        HubBolt.Config validationHubConfig = HubBolt.Config.builder()
+                .requestSenderComponent(ROUTER_BOLT_NAME)
+                .workerComponent(VALIDATION_WORKER_BOLT)
+                .timeoutMs((int) TimeUnit.SECONDS.toMillis(topologyConfig.getProcessTimeout()))
+                .build();
+        tb.setBolt(FlowValidationHubBolt.ID,
+                new FlowValidationHubBolt(validationHubConfig, persistenceManager, flowResourcesConfig,
+                        topologyConfig.getFlowMeterMinBurstSizeInKbits(),
+                        topologyConfig.getFlowMeterBurstCoefficient()))
+                .fieldsGrouping(ROUTER_BOLT_NAME, FlowValidationHubBolt.INCOME_STREAM, FIELDS_KEY)
+                .directGrouping(VALIDATION_WORKER_BOLT, FlowValidationHubBolt.INCOME_STREAM)
+                .directGrouping(CoordinatorBolt.ID);
+
+        WorkerBolt.Config speakerValidationWorkerConfig = WorkerBolt.Config.builder()
+                .hubComponent(FlowValidationHubBolt.ID)
+                .streamToHub(FlowValidationHubBolt.INCOME_STREAM)
+                .workerSpoutComponent(ROUTER_BOLT_NAME)
+                .defaultTimeout((int) TimeUnit.SECONDS.toMillis(topologyConfig.getOperationTimeout()))
+                .build();
+        tb.setBolt(VALIDATION_WORKER_BOLT, new SpeakerWorkerBolt(speakerValidationWorkerConfig))
+                .fieldsGrouping(ROUTER_BOLT_NAME, SpeakerWorkerBolt.INCOME_STREAM, FIELDS_KEY)
+                .fieldsGrouping(FlowValidationHubBolt.ID, StreamType.FLOW_VALIDATION_WORKER.toString(), FIELDS_KEY)
+                .directGrouping(CoordinatorBolt.ID);
+
+        HubBolt.Config meterModifyHubConfig = HubBolt.Config.builder()
+                .requestSenderComponent(ROUTER_BOLT_NAME)
+                .workerComponent(METER_MODIFY_WORKER_BOLT)
+                .timeoutMs((int) TimeUnit.SECONDS.toMillis(topologyConfig.getProcessTimeout()))
+                .build();
+        tb.setBolt(FlowMeterModifyHubBolt.ID, new FlowMeterModifyHubBolt(meterModifyHubConfig, persistenceManager))
+                .fieldsGrouping(ROUTER_BOLT_NAME, FlowMeterModifyHubBolt.INCOME_STREAM, FIELDS_KEY)
+                .directGrouping(METER_MODIFY_WORKER_BOLT, FlowMeterModifyHubBolt.INCOME_STREAM)
+                .directGrouping(CoordinatorBolt.ID);
+
+        WorkerBolt.Config speakerMeterModifyWorkerConfig = WorkerBolt.Config.builder()
+                .hubComponent(FlowMeterModifyHubBolt.ID)
+                .streamToHub(FlowMeterModifyHubBolt.INCOME_STREAM)
+                .workerSpoutComponent(ROUTER_BOLT_NAME)
+                .defaultTimeout((int) TimeUnit.SECONDS.toMillis(topologyConfig.getOperationTimeout()))
+                .build();
+        tb.setBolt(METER_MODIFY_WORKER_BOLT, new SpeakerWorkerBolt(speakerMeterModifyWorkerConfig))
+                .fieldsGrouping(ROUTER_BOLT_NAME, SpeakerWorkerBolt.INCOME_STREAM, FIELDS_KEY)
+                .fieldsGrouping(FlowMeterModifyHubBolt.ID, StreamType.METER_MODIFY_WORKER.toString(), FIELDS_KEY)
+                .directGrouping(CoordinatorBolt.ID);
 
         SwitchOperationsBolt switchesBolt = new SwitchOperationsBolt(persistenceManager);
         tb.setBolt(SWITCHES_BOLT_NAME, switchesBolt, parallelism)
@@ -121,10 +190,6 @@ public class NbWorkerTopology extends AbstractTopology<NbWorkerTopologyConfig> {
         tb.setBolt(PATHS_BOLT_NAME, pathsBolt, parallelism)
                 .shuffleGrouping(ROUTER_BOLT_NAME, StreamType.PATHS.toString());
 
-        SwitchValidationsBolt validationBolt = new SwitchValidationsBolt(persistenceManager);
-        tb.setBolt(SWITCH_VALIDATIONS_BOLT_NAME, validationBolt, parallelism)
-                .shuffleGrouping(ROUTER_BOLT_NAME, StreamType.VALIDATION.toString());
-
         HistoryOperationsBolt historyBolt = new HistoryOperationsBolt(persistenceManager);
         tb.setBolt(HISTORY_BOLT_NAME, historyBolt, parallelism)
                 .shuffleGrouping(ROUTER_BOLT_NAME, StreamType.HISTORY.toString());
@@ -137,8 +202,9 @@ public class NbWorkerTopology extends AbstractTopology<NbWorkerTopologyConfig> {
                 .shuffleGrouping(FEATURE_TOGGLES_BOLT_NAME)
                 .shuffleGrouping(KILDA_CONFIG_BOLT_NAME)
                 .shuffleGrouping(PATHS_BOLT_NAME)
-                .shuffleGrouping(SWITCH_VALIDATIONS_BOLT_NAME)
-                .shuffleGrouping(HISTORY_BOLT_NAME);
+                .shuffleGrouping(HISTORY_BOLT_NAME)
+                .shuffleGrouping(FlowValidationHubBolt.ID)
+                .shuffleGrouping(FlowMeterModifyHubBolt.ID);
 
         MessageEncoder messageEncoder = new MessageEncoder();
         tb.setBolt(MESSAGE_ENCODER_BOLT_NAME, messageEncoder, parallelism)
@@ -155,7 +221,9 @@ public class NbWorkerTopology extends AbstractTopology<NbWorkerTopologyConfig> {
                 .shuffleGrouping(FEATURE_TOGGLES_BOLT_NAME, StreamType.ERROR.toString())
                 .shuffleGrouping(KILDA_CONFIG_BOLT_NAME, StreamType.ERROR.toString())
                 .shuffleGrouping(PATHS_BOLT_NAME, StreamType.ERROR.toString())
-                .shuffleGrouping(HISTORY_BOLT_NAME, StreamType.ERROR.toString());
+                .shuffleGrouping(HISTORY_BOLT_NAME, StreamType.ERROR.toString())
+                .shuffleGrouping(FlowValidationHubBolt.ID, StreamType.ERROR.toString())
+                .shuffleGrouping(FlowMeterModifyHubBolt.ID, StreamType.ERROR.toString());
 
         DiscoveryEncoderBolt discoveryEncoder = new DiscoveryEncoderBolt();
         tb.setBolt(DISCOVERY_ENCODER_BOLT_NAME, discoveryEncoder, parallelism)
@@ -179,6 +247,10 @@ public class NbWorkerTopology extends AbstractTopology<NbWorkerTopologyConfig> {
         KafkaBolt kafkaDiscoBolt = buildKafkaBolt(topologyConfig.getKafkaDiscoTopic());
         tb.setBolt(DISCO_KAFKA_BOLT_NAME, kafkaDiscoBolt, parallelism)
                 .shuffleGrouping(DISCOVERY_ENCODER_BOLT_NAME);
+
+        tb.setBolt(SPEAKER_KAFKA_BOLT, buildKafkaBolt(topologyConfig.getKafkaSpeakerTopic()))
+                .shuffleGrouping(VALIDATION_WORKER_BOLT, StreamType.TO_SPEAKER.toString())
+                .shuffleGrouping(METER_MODIFY_WORKER_BOLT, StreamType.TO_SPEAKER.toString());
 
         return tb.createTopology();
     }